perl drafting.

+++ perl/index.pl
@@ -21,13 +21,17 @@ binmode( STDIN,  ':encoding(utf8)' );
 binmode( STDOUT, ':encoding(utf8)' );
 binmode( STDERR, ':encoding(utf8)' );
 
-my $mydir=dirname(abs_path($0));
-my $basename=basename($0,(".pl"));
+my $mydir = dirname(abs_path($0));
+my $basename = basename($0,(".pl"));
+my $singlerun = 1;
 chdir($mydir);
 print "workdir:[".$mydir."]\tbasename:[".$basename."]\n";
-
-open(LOCK,">".$mydir."/".$basename.".lock") || die $!;
-flock(LOCK,LOCK_EX|LOCK_NB) || warn "another $basename is running,exit\n";
+if($singlerun == 1){
+	open(LOCK,">".$mydir."/".$basename.".lock") || die $!;
+	flock(LOCK,LOCK_EX|LOCK_NB) || warn "another $basename is running,exit\n";
+}
+use lib '/mnt/hgfs/HostGitHub/GWA2/perl'; # @todo
+use lib './';
 
 # main body
 
@@ -35,19 +39,27 @@ flock(LOCK,LOCK_EX|LOCK_NB) || warn "another $basename is running,exit\n";
 
 print "now is: ".strftime("%Y-%m-%d-%H:%M:%S", localtime)."\n";
 
-use lib '/mnt/hgfs/HostGitHub/GWA2/perl';
-use lib './';
-
-my $_ctrl_var_a = "I am a var from $0, 1:[".$ARGV[0]."]";
+my $i = 100;
+my $hmf;
+my %hmf = ();
+my $_ctrl_var_a = "I:[$i] am a var from $0, 1:[".$ARGV[0]."] index.";
 my $ctrl = "index";
+my $argvsize = @ARGV;
+
+$hmf{'var_a'} = $_ctrl_var_a;
+$ARGV[$argvsize+1] = $hmf;
+
+print "bfr var-a:[$_ctrl_var_a] in index 1:[".$ARGV[1]."] size:[".@ARGV."] 3:[".$ARGV[3]."].\n";
 
 require "./ctrl/$ctrl.pl";
 
-print "var-a:[$_ctrl_var_a] in index 1:[".$ARGV[1]."].\n";
+print "aft var-a:[$_ctrl_var_a] in index 1:[".$ARGV[1]."] 2:[".$ARGV[2]."].\n";
 
 #__exec__($ARGV); # why?
 
 # comm/footer.inc.pl ?
 
-close(LOCK);
-unlink($mydir."/".$basename.".lock") or warn "remove lock file failed:[$!].\n";
+if($singlerun == 1){
+	close(LOCK);
+	unlink($mydir."/".$basename.".lock") or warn "remove lock file failed:[$!].\n";
+}
